feat: theme and database wrappers
