+++ frontend/wrappers/DatabaseWrapper.tsx
@@ -1,0 +1,14 @@
+import { db, DB_Type } from "@/database/db";
+import { DbLocationHook } from "@pages/hooks/dbLocationHook";
+import { useEffect } from "react";
+
+export default function DatabaseWrapper({ children }: { children: JSX.Element }) {
+  const { changeDbLocation } = DbLocationHook();
+
+  useEffect(() => {
+    !db().getDbLocation() && db().setDbLocation(DB_Type.LOCAL);
+    changeDbLocation(db().getDbLocation() || DB_Type.LOCAL);
+  }, []);
+
+  return <>{children}</>;
+}
